Readers and README update
